+++ provider/plugin/oauth/controller/application/create_test.go
@@ -7,9 +7,9 @@ import (
 	"net/http"
 	"testing"
 
-	"github.com/codefluence-x/altair/entity"
-	"github.com/codefluence-x/altair/eobject"
 	"github.com/codefluence-x/altair/provider/plugin/oauth/controller"
+	"github.com/codefluence-x/altair/provider/plugin/oauth/entity"
+	"github.com/codefluence-x/altair/provider/plugin/oauth/eobject"
 	"github.com/codefluence-x/altair/provider/plugin/oauth/mock"
 	"github.com/codefluence-x/altair/testhelper"
 	"github.com/codefluence-x/altair/util"
@@ -17,6 +17,10 @@ import (
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/assert"
 )
+
+type ErrorResponse struct {
+	Errors []entity.ErrorObject `json:"errors"`
+}
 
 func TestCreate(t *testing.T) {
 	mockCtrl := gomock.NewController(t)
@@ -87,7 +91,7 @@ func TestCreate(t *testing.T) {
 				ctrl := controller.Application().Create(applicationManager)
 				apiEngine.Handle(ctrl.Method(), ctrl.Path(), ctrl.Control)
 
-				var response testhelper.ErrorResponse
+				var response ErrorResponse
 				w := testhelper.PerformRequest(apiEngine, ctrl.Method(), ctrl.Path(), bytes.NewReader(encodedBytes))
 
 				err = json.Unmarshal(w.Body.Bytes(), &response)
@@ -113,7 +117,7 @@ func TestCreate(t *testing.T) {
 					Errors:     eobject.Wrap(eobject.BadRequestError("request body")),
 				}
 
-				var response testhelper.ErrorResponse
+				var response ErrorResponse
 				w := testhelper.PerformRequest(apiEngine, ctrl.Method(), ctrl.Path(), testhelper.MockErrorIoReader{})
 
 				err := json.Unmarshal(w.Body.Bytes(), &response)
@@ -139,7 +143,7 @@ func TestCreate(t *testing.T) {
 					Errors:     eobject.Wrap(eobject.BadRequestError("request body")),
 				}
 
-				var response testhelper.ErrorResponse
+				var response ErrorResponse
 				w := testhelper.PerformRequest(apiEngine, ctrl.Method(), ctrl.Path(), bytes.NewReader([]byte(`this is gonna be error`)))
 
 				err := json.Unmarshal(w.Body.Bytes(), &response)
