+++ provider/plugin/oauth/controller/application/one.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 	"strconv"
 
-	"github.com/codefluence-x/altair/eobject"
+	"github.com/codefluence-x/altair/provider/plugin/oauth/eobject"
 	"github.com/codefluence-x/altair/provider/plugin/oauth/interfaces"
 	"github.com/codefluence-x/journal"
 	"github.com/gin-gonic/gin"
