+++ provider/plugin/oauth/controller/authorization/revoke_test.go
@@ -6,8 +6,8 @@ import (
 	"net/http"
 	"testing"
 
-	"github.com/codefluence-x/altair/entity"
-	"github.com/codefluence-x/altair/eobject"
+	"github.com/codefluence-x/altair/provider/plugin/oauth/entity"
+	"github.com/codefluence-x/altair/provider/plugin/oauth/eobject"
 
 	"github.com/codefluence-x/altair/testhelper"
 
@@ -19,6 +19,10 @@ import (
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/assert"
 )
+
+type ErrorResponse struct {
+	Errors []entity.ErrorObject `json:"errors"`
+}
 
 func TestRevoke(t *testing.T) {
 	mockCtrl := gomock.NewController(t)
@@ -85,7 +89,7 @@ func TestRevoke(t *testing.T) {
 					ctrl := controller.Authorization().Revoke(authorizationService)
 					apiEngine.Handle(ctrl.Method(), ctrl.Path(), ctrl.Control)
 
-					var response testhelper.ErrorResponse
+					var response ErrorResponse
 					w := testhelper.PerformRequest(apiEngine, ctrl.Method(), ctrl.Path(), bytes.NewReader(encodedBytes))
 
 					err = json.Unmarshal(w.Body.Bytes(), &response)
@@ -112,7 +116,7 @@ func TestRevoke(t *testing.T) {
 					Errors:     eobject.Wrap(eobject.BadRequestError("request body")),
 				}
 
-				var response testhelper.ErrorResponse
+				var response ErrorResponse
 				w := testhelper.PerformRequest(apiEngine, ctrl.Method(), ctrl.Path(), testhelper.MockErrorIoReader{})
 
 				err := json.Unmarshal(w.Body.Bytes(), &response)
@@ -138,7 +142,7 @@ func TestRevoke(t *testing.T) {
 					Errors:     eobject.Wrap(eobject.BadRequestError("request body")),
 				}
 
-				var response testhelper.ErrorResponse
+				var response ErrorResponse
 				w := testhelper.PerformRequest(apiEngine, ctrl.Method(), ctrl.Path(), bytes.NewReader([]byte(`this is gonna be error`)))
 
 				err := json.Unmarshal(w.Body.Bytes(), &response)
