+++ provider/plugin/oauth/formatter/oauth_application.go
@@ -3,7 +3,7 @@ package formatter
 import (
 	"context"
 
-	"github.com/codefluence-x/altair/entity"
+	"github.com/codefluence-x/altair/provider/plugin/oauth/entity"
 	"github.com/codefluence-x/altair/util"
 )
