+++ provider/plugin/oauth/formatter/oauth_formatter_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/codefluence-x/altair/entity"
+	"github.com/codefluence-x/altair/provider/plugin/oauth/entity"
 	"github.com/codefluence-x/altair/provider/plugin/oauth/formatter"
 
 	"github.com/codefluence-x/altair/util"
