+++ provider/plugin/oauth/model/oauth_access_token.go
@@ -5,7 +5,7 @@ import (
 	"database/sql"
 	"time"
 
-	"github.com/codefluence-x/altair/entity"
+	"github.com/codefluence-x/altair/provider/plugin/oauth/entity"
 	"github.com/codefluence-x/altair/provider/plugin/oauth/query"
 )
