+++ provider/plugin/oauth/oauth.go
@@ -4,9 +4,9 @@ import (
 	"time"
 
 	"github.com/codefluence-x/altair/core"
-	"github.com/codefluence-x/altair/entity"
 	"github.com/codefluence-x/altair/provider/plugin/oauth/controller"
 	"github.com/codefluence-x/altair/provider/plugin/oauth/downstream"
+	"github.com/codefluence-x/altair/provider/plugin/oauth/entity"
 	"github.com/codefluence-x/altair/provider/plugin/oauth/formatter"
 	"github.com/codefluence-x/altair/provider/plugin/oauth/model"
 	"github.com/codefluence-x/altair/provider/plugin/oauth/service"
